Separate some common trakt functions to separate file
